pchat: My instant messaging client
I'm fed up with the state of instant messaging clients so I've decided
to attempt to write my own.  I'm hoping that this will eventually
share a whole bunch of code with MHng, with the goal of eventually
splitting some stuff off into libraries and defining a sane IPC
mechanism for these sorts of applications.